+++ lib/src/extended_image.dart
@@ -996,8 +996,8 @@ class _ExtendedImageState extends State<ExtendedImage>
   @override
   void dispose() {
     assert(_imageStream != null);
-
-    WidgetsBinding.instance.removeObserver(this);
+    // ignore: unnecessary_non_null_assertion
+    WidgetsBinding.instance!.removeObserver(this);
     _stopListeningToStream();
     _completerHandle?.dispose();
     _scrollAwareContext.dispose();
@@ -1026,7 +1026,8 @@ class _ExtendedImageState extends State<ExtendedImage>
     super.initState();
     returnLoadStateChangedWidget = false;
     _loadState = LoadState.loading;
-    WidgetsBinding.instance.addObserver(this);
+    // ignore: unnecessary_non_null_assertion
+    WidgetsBinding.instance!.addObserver(this);
     _scrollAwareContext = DisposableBuildContext<State<ExtendedImage>>(this);
   }
 
@@ -1208,7 +1209,8 @@ class _ExtendedImageState extends State<ExtendedImage>
 
   void _updateInvertColors() {
     _invertColors = MediaQuery.maybeOf(context)?.invertColors ??
-        SemanticsBinding.instance.accessibilityFeatures.invertColors;
+        // ignore: unnecessary_non_null_assertion
+        SemanticsBinding.instance!.accessibilityFeatures.invertColors;
   }
 
   void _updateSourceStream(ImageStream newStream, {bool rebuild = false}) {